requirements.txt
streamlit==1.28.0
streamlit-drawable-canvas==0.9.3
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
sqlalchemy
psycopg[binary]>=3.2.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
pandas>=2.0.0
reportlab>=4.0.0
openai>=1.0.0
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0
ultralytics>=8.0.0
